Move shuffle_extrinsics() to utils module
