+++ domains/client/domain-executor/src/core_bundle_processor.rs
@@ -1,11 +1,9 @@
 use crate::fraud_proof::{find_trace_mismatch, FraudProofGenerator};
+use crate::utils::shuffle_extrinsics;
 use crate::TransactionFor;
 use codec::{Decode, Encode};
 use domain_block_builder::{BlockBuilder, BuiltBlock, RecordProof};
 use domain_runtime_primitives::{AccountId, DomainCoreApi};
-use rand::seq::SliceRandom;
-use rand::SeedableRng;
-use rand_chacha::ChaCha8Rng;
 use sc_client_api::{AuxStore, BlockBackend};
 use sc_consensus::{
     BlockImport, BlockImportParams, ForkChoiceStrategy, ImportResult, StateAction, StorageChanges,
@@ -23,65 +21,12 @@ use sp_runtime::generic::BlockId;
 use sp_runtime::traits::{Block as BlockT, HashFor, Header as HeaderT, One};
 use sp_runtime::Digest;
 use std::borrow::Cow;
-use std::collections::{BTreeMap, VecDeque};
-use std::fmt::Debug;
 use std::marker::PhantomData;
 use std::sync::Arc;
 use subspace_core_primitives::{BlockNumber, Randomness};
 use system_runtime_primitives::SystemDomainApi;
 
 const LOG_TARGET: &str = "bundle-processor";
-
-/// Shuffles the extrinsics in a deterministic way.
-///
-/// The extrinsics are grouped by the signer. The extrinsics without a signer, i.e., unsigned
-/// extrinsics, are considered as a special group. The items in different groups are cross shuffled,
-/// while the order of items inside the same group is still maintained.
-fn shuffle_extrinsics<Extrinsic: Debug>(
-    extrinsics: Vec<(Option<AccountId>, Extrinsic)>,
-    shuffling_seed: Randomness,
-) -> Vec<Extrinsic> {
-    let mut rng = ChaCha8Rng::from_seed(shuffling_seed);
-
-    let mut positions = extrinsics
-        .iter()
-        .map(|(maybe_signer, _)| maybe_signer)
-        .cloned()
-        .collect::<Vec<_>>();
-
-    // Shuffles the positions using Fisher–Yates algorithm.
-    positions.shuffle(&mut rng);
-
-    let mut grouped_extrinsics: BTreeMap<Option<AccountId>, VecDeque<_>> = extrinsics
-        .into_iter()
-        .fold(BTreeMap::new(), |mut groups, (maybe_signer, tx)| {
-            groups
-                .entry(maybe_signer)
-                .or_insert_with(VecDeque::new)
-                .push_back(tx);
-            groups
-        });
-
-    // The relative ordering for the items in the same group does not change.
-    let shuffled_extrinsics = positions
-        .into_iter()
-        .map(|maybe_signer| {
-            grouped_extrinsics
-                .get_mut(&maybe_signer)
-                .expect("Extrinsics are grouped correctly; qed")
-                .pop_front()
-                .expect("Extrinsic definitely exists as it's correctly grouped above; qed")
-        })
-        .collect::<Vec<_>>();
-
-    tracing::trace!(
-        target: LOG_TARGET,
-        ?shuffled_extrinsics,
-        "Shuffled extrinsics"
-    );
-
-    shuffled_extrinsics
-}
 
 pub(crate) struct CoreBundleProcessor<Block, SBlock, PBlock, Client, SClient, PClient, Backend, E>
 where
@@ -635,40 +580,5 @@ where
         }
 
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use sp_keyring::sr25519::Keyring;
-    use sp_runtime::traits::{BlakeTwo256, Hash as HashT};
-
-    #[test]
-    fn shuffle_extrinsics_should_work() {
-        let alice = Keyring::Alice.to_account_id();
-        let bob = Keyring::Bob.to_account_id();
-        let charlie = Keyring::Charlie.to_account_id();
-
-        let extrinsics = vec![
-            (Some(alice.clone()), 10),
-            (None, 100),
-            (Some(bob.clone()), 1),
-            (Some(bob), 2),
-            (Some(charlie.clone()), 30),
-            (Some(alice.clone()), 11),
-            (Some(charlie), 31),
-            (None, 101),
-            (None, 102),
-            (Some(alice), 12),
-        ];
-
-        let dummy_seed = BlakeTwo256::hash_of(&[1u8; 64]).into();
-        let shuffled_extrinsics = shuffle_extrinsics(extrinsics, dummy_seed);
-
-        assert_eq!(
-            shuffled_extrinsics,
-            vec![100, 30, 10, 1, 11, 101, 31, 12, 102, 2]
-        );
     }
 }
